Task 10 Finished

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -11,7 +11,8 @@
 
 size_t print_listint(const listint_t *h)
 {
-	size_t n = 0;
+	unsigned int n = 0;
+	
 
 	while (h)
 	{
